requirements.txt
fastapi
uvicorn[standard]
python-socketio
eventlet
openai
python-dotenv
python-decouple
pyjwt
pymongo
